core: riscv: Regard supervisor software interrupt as foreign interrupt
Refer to OP-TEE issue #6767 [1], Linux generates IPI as software
interrupt to other harts, even if the target hart is in OP-TEE world.
The hart in OP-TEE which receives supervisor software interrupt should
yield the execution back to Linux to handle that software interrupt.

In this commit, we classify CSR_XIE_SIE into THREAD_EXCP_FOREIGN_INTR,
and call thread_foreign_interrupt_handler() when a software interrupt is
taken to issue RPC to Linux.

[1]: https://github.com/OP-TEE/optee_os/issues/6767

Signed-off-by: Alvin Chang <alvinga@andestech.com>